requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
alembic
mysqlclient   # OR use 'pymysql'
passlib[bcrypt]
